添加逾期查询接口

+++ src/main/java/com/yusjade/libms/service/BookService.java
@@ -1,7 +1,9 @@
 package com.yusjade.libms.service;
 
+import com.yusjade.libms.dao.BookInventoryMapper;
 import com.yusjade.libms.dao.BookMapper;
 import com.yusjade.libms.pojo.Book;
+import com.yusjade.libms.pojo.BookInventory;
 import jakarta.annotation.Resource;
 import java.lang.management.LockInfo;
 import java.util.List;
@@ -12,6 +14,8 @@ public class BookService implements BaseService<Book, Book> {
 
   @Resource
   BookMapper bookMapper;
+  @Resource
+  BookInventoryMapper bookInventoryMapper;
 
   public List<Book> listBookSelective(Long bookId, Long inventoryId, Boolean isBorrowed,
       Boolean isDiscarded) {
@@ -29,6 +33,16 @@ public class BookService implements BaseService<Book, Book> {
   }
 
   public Integer removeBook(Long id) {
+    Book record = bookMapper.selectByPrimaryKey(id);
+    if (record == null) {
+      return 0;
+    }
+    if (record.getInventoryId() != null) {
+      // 更新书库数量
+      BookInventory inventory = bookInventoryMapper.selectByPrimaryKey(record.getInventoryId());
+      inventory.setQuantity(inventory.getQuantity() - 1);
+      bookInventoryMapper.updateByPrimaryKeySelective(inventory);
+    }
     return bookMapper.deleteByPrimaryKey(id);
   }
 
@@ -37,6 +51,10 @@ public class BookService implements BaseService<Book, Book> {
     record.setIsBorrowed(false);
     record.setIsDiscarded(false);
     if (bookMapper.insert(record) == 1) {
+      // 更新书库数量
+      BookInventory inventory = bookInventoryMapper.selectByPrimaryKey(record.getInventoryId());
+      inventory.setQuantity(inventory.getQuantity() + 1);
+      bookInventoryMapper.updateByPrimaryKeySelective(inventory);
       return record.getBookId();
     }
     return 0L;
